axe: fix axe_get_signal()

+++ src/axe.c
@@ -669,14 +669,14 @@ fe_delivery_system_t axe_delsys(int aid, int fd, fe_delivery_system_t *sys)
 
 void axe_get_signal(adapter *ad)
 {
-	uint16_t strength = 0, snr = 0;
-	uint32_t status = 0, ber = 0, tmp;
+	int strength = 0, snr = 0;
+	int status = 0, ber = 0, tmp;
 	get_signal(ad, &status, &ber, &strength, &snr);
 
 	strength = strength * 240 / 24000;
 	if (strength > 240)
 		strength = 240;
-	tmp = (uint32_t)snr * 255 / 54000;
+	tmp = snr * 255 / 54000;
 	if (tmp > 255)
 		tmp = 255;
 	if (tmp <= 15)
